Separate Pubid::Nist::Identifier and Pubid::Nist::Identifier::Base as implemented in other pubid-* gems
