+++ spec/nist_pubid/document/nbs_hb_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Pubid::Nist::Identifier do
+RSpec.describe Pubid::Nist::Identifier::Base do
   subject { described_class.parse(original_pubid) }
 
   let(:mr_pubid) { short_pubid.gsub(" ", ".") }
